requirements.txt
dialoguekit
ollama
python-dotenv